% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/altarr.R
\name{altarr_reset}
\alias{altarr_reset}
\title{Reset counters, and optionally drop the chunk cache}
\usage{
altarr_reset(x, cache = TRUE)
}
\arguments{
\item{x}{an altarr array}

\item{cache}{drop cached chunks too?}
}
\description{
Reset counters, and optionally drop the chunk cache
}
